+++ src/logger/logger.midleware.ts
@@ -59,9 +59,7 @@ export class LoggerMiddleware implements NestMiddleware {
     const formattedStartTime = new Date(startTime).toISOString();
     const formattedEndTime = new Date(endTime).toISOString();
 
-    return `${method} ${url} ${JSON.stringify(
-      query,
-    )} ${JSON.stringify(
+    return `${method} ${url} ${JSON.stringify(query)} ${JSON.stringify(
       body,
     )} ${statusCode} ${contentLength}B - ${userAgent} - ${formattedStartTime} - ${formattedEndTime} - ${elapsedTime}ms`;
   }
