+++ src/utils/common.ts
@@ -1,0 +1,6 @@
+import { compare } from "bcrypt";
+
+export const isCorrectPassword = async (
+  oldPassword: string,
+  currentPassword: string,
+) => await compare(oldPassword, currentPassword);
