+++ src/utils/types.ts
@@ -1,5 +1,3 @@
-import { compare } from "bcrypt";
-
 export interface User {
   id?: string;
   login: string;
@@ -35,10 +33,4 @@ export interface Favorites {
   albums: string[];
   tracks: string[];
 }
-
-export const isCorrectPassword = async (
-  oldPassword: string,
-  currentPassword: string,
-) => await compare(oldPassword, currentPassword);
-
